+++ src/Telegram.Bot/Types/InlineQueryResults/InlineQueryResultPhoto.cs
@@ -54,12 +54,9 @@ namespace Telegram.Bot.Types.InlineQueryResults
 
         /// <inheritdoc />
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public InputMessageContent InputMessageContent { get; set; }
+        public InputMessageContentBase InputMessageContent { get; set; }
 
-        /// <summary>
-        /// Initializes a new inline query result
-        /// </summary>
-        public InlineQueryResultPhoto()
+        private InlineQueryResultPhoto()
             : base(InlineQueryResultType.Photo)
         {
         }
@@ -71,9 +68,8 @@ namespace Telegram.Bot.Types.InlineQueryResults
         /// <param name="photoUrl">A valid URL of the photo. Photo size must not exceed 5MB.</param>
         /// <param name="thumbUrl">Optional. Url of the thumbnail for the result.</param>
         public InlineQueryResultPhoto(string id, string photoUrl, string thumbUrl)
-            : this()
+            : base(InlineQueryResultType.Photo, id)
         {
-            Id = id;
             PhotoUrl = photoUrl;
             ThumbUrl = thumbUrl;
         }