+++ src/Telegram.Bot/Types/InlineQueryResults/InlineQueryResultVoice.cs
@@ -38,15 +38,14 @@ namespace Telegram.Bot.Types.InlineQueryResults
 
         /// <inheritdoc />
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public InputMessageContent InputMessageContent { get; set; }
+        public InputMessageContentBase InputMessageContent { get; set; }
 
         /// <summary>
         /// Initializes a new inline query result
         /// </summary>
         public InlineQueryResultVoice()
             : base(InlineQueryResultType.Voice)
-        {
-        }
+        { }
 
         /// <summary>
         /// Initializes a new inline query result
@@ -55,9 +54,8 @@ namespace Telegram.Bot.Types.InlineQueryResults
         /// <param name="voiceUrl">A valid URL for the voice recording</param>
         /// <param name="title">Title of the result</param>
         public InlineQueryResultVoice(string id, string voiceUrl, string title)
-            : this()
+            : base(InlineQueryResultType.Voice, id)
         {
-            Id = id;
             VoiceUrl = voiceUrl;
             Title = title;
         }